Add layout and migrate home view to layout without children
