+++ src/app/layout.js
@@ -3,6 +3,7 @@ import { useReducer } from "react";
 import MeliContext from "@/store/meliContext";
 import meliReducer from "@/store/meliReducer";
 import initialMeliState from "@/store/initialMeliState";
+import MeliLayout from "@/coreComponents/MeliLayout";
 import { Inter } from "next/font/google";
 import "./globals.css";
 
@@ -14,7 +15,7 @@ export default function RootLayout({ children }) {
     <html lang="en">
       <body className={inter.className}>
         <MeliContext.Provider value={[state, dispatch]}>
-          {children}
+          <MeliLayout>{children}</MeliLayout>
         </MeliContext.Provider>
       </body>
     </html>
