+++ src/coreComponents/InputSearch/InputSearch.module.scss
@@ -2,10 +2,9 @@
   display: flex;
   flex-direction: row;
   align-items: center;
-  border: 1px solid #eeeeee;
   width: 80%;
   input {
-    height: 3em;
+    height: 45px;
     border: 0;
     width: 95%;
     padding: 0em 1em;
@@ -30,6 +29,6 @@
     width: 5%;
     height: 100%;
     cursor: pointer;
-    padding: 1.2em 0em;
+    padding: 11px 0em;
   }
 }
