+++ src/coreComponents/MeliLayout/meliLayout.module.scss
@@ -1,0 +1,36 @@
+.layoutContainer {
+  height: 100dvh;
+  .layoutHead {
+    width: 100%;
+    height: 5dvh;
+    background-color: #ffe600;
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    .searchContainer {
+      display: flex;
+      flex-direction: row;
+      justify-content: center;
+      align-items: center;
+      width: 70%;
+      .mlLogo {
+        margin-right: 2em;
+      }
+    }
+  }
+  .sectionContainer {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    height: 95dvh;
+    background-color: #eeeeee;
+    .bodyContainer {
+      width: 60%;
+      height: 85dvh;
+      background-color: #ffffff;
+    }
+  }
+}
